+++ src/main/kotlin/io/redgreen/timelapse/visuals/FormattedDiff.kt
@@ -5,12 +5,19 @@ import io.redgreen.timelapse.visuals.DiffSpan.Deletion
 import io.redgreen.timelapse.visuals.DiffSpan.Insertion
 import io.redgreen.timelapse.visuals.DiffSpan.Unmodified
 
+private const val DIFF_TYPE_DELETED = "deleted"
+private const val DIFF_TYPE_NEW_FILE = "new file"
+
 class FormattedDiff(val spans: List<DiffSpan>) {
   companion object {
     fun from(rawDiff: String): FormattedDiff {
+      val diffTypeLine = rawDiff.lines().take(2).last()
+      val deletedOrNewFile = with(diffTypeLine) { startsWith(DIFF_TYPE_NEW_FILE) || startsWith(DIFF_TYPE_DELETED) }
+      val skipLines = if (deletedOrNewFile) { 5 } else { 4 }
+
       val spans = rawDiff
         .lines()
-        .filterIndexed { index, _ -> index > 4 }
+        .filterIndexed { index, _ -> index > skipLines }
         .map(::toSpan)
 
       return FormattedDiff(spans)