+++ src/test/kotlin/io/redgreen/timelapse/visuals/FormattedDiffTest.kt
@@ -41,4 +41,54 @@ class FormattedDiffTest {
         Unmodified("    mavenCentral()"),
       )
   }
+
+  @Test
+  fun `it should create a formatted diff for a deleted file`() {
+    // given
+    val rawDiff = """
+      diff --git a/file-4.txt b/file-4.txt
+      deleted file mode 100644
+      index 980a0d5..0000000
+      --- a/file-4.txt
+      +++ /dev/null
+      @@ -1 +0,0 @@
+      -Hello World!
+       
+    """.trimIndent()
+
+    // when
+    val formattedDiff = FormattedDiff.from(rawDiff)
+
+    // then
+    assertThat(formattedDiff.spans)
+      .containsExactly(
+        Deletion("Hello World!"),
+        Blank,
+      )
+  }
+
+  @Test
+  fun `it should create a formatted diff for a newly added file`() {
+    // given
+    val rawDiff = """
+      diff --git a/file-c.txt b/file-c.txt
+      new file mode 100644
+      index 0000000..8bc25bf
+      --- /dev/null
+      +++ b/file-c.txt
+      @@ -0,0 +1 @@
+      +I'm new in town :)
+       
+    """.trimIndent()
+
+    // when
+    val formattedDiff = FormattedDiff.from(rawDiff)
+
+    // then
+    assertThat(formattedDiff.spans)
+      .containsExactly(
+        Insertion("I'm new in town :)"),
+        Blank,
+      )
+  }
 }